move stateNodeTutorial
